SaaS/app-aviladevops/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
gunicorn==23.0.0
whitenoise==6.11.0
dj-database-url==2.2.0
python-dotenv==1.0.1